requirements.txt
streamlit
pandas
plotly
gspread
oauth2client
python-telegram-bot